Adding fixes from Philippe Marschall
I attached two patches.

KomParseCookiesFix
This additionally parses cookies that have only a key but no value.
Getting this in will allow us to remove an override and help Pavel with the KernelImage.

KomSqueak39Fix
Gives Kom in Squeak 3.9 the same semantics as in Squeak 3.8. Getting this in will make WAKom work on Squeak 3.9 and allow us to remove WAKomEncoded39. I have not checked if it works on Squeak 3.8 maybe you need to make a Kom for Squeak 3.8 and one for 3.9.

Co-authored-by: Philippe Marshall <root@gmail.com>
Co-authored-by: Philippe Marshall <root@gmail.com>
Co-authored-by: Philippe Marshall <root@gmail.com>